chore: resolve code duplicate issue
